# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LumenBazaar Backend API",
    "version": "0.1.0",
    "description": "Facilitator, seller catalog, discovery, and conformance APIs for LumenBazaar."
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Local development" }],
...
  "paths": {
...
    "/v1/supported": {
      "get": {
        "operationId": "listSupportedPaymentSchemes",
        "summary": "List supported x402 payment schemes.",
        "tags": ["facilitator"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["kinds", "extensions", "signers"],
                  "properties": {
                    "kinds": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["x402Version", "scheme", "network"],
                        "properties": {
                          "x402Version": { "type": "integer", "enum": [2] },
                          "scheme": { "type": "string" },
                          "network": { "type": "string" },
                          "extra": { "type": "object", "additionalProperties": true }
                        }
                      }
                    },
                    "extensions": { "type": "array", "items": { "type": "string" } },
                    "signers": {
                      "type": "object",
                      "additionalProperties": { "type": "array", "items": { "type": "string" } }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok", "error"],
                  "properties": {
                    "ok": { "type": "boolean", "enum": [false] },
                    "error": {
                      "type": "object",
                      "required": ["code", "message"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "details": { "type": "object", "additionalProperties": true }
                      }
                    },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/verify": {
      "post": {
        "operationId": "verifyPayment",
        "summary": "Verify an exact Stellar x402 payment payload.",
        "tags": ["facilitator"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["x402Version", "paymentPayload", "paymentRequirements"],
                "properties": {
                  "x402Version": { "type": "integer", "enum": [2] },
                  "paymentPayload": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": ["x402Version", "accepted", "payload"],
                    "properties": {
                      "x402Version": { "type": "integer", "enum": [2] },
                      "resource": {
                        "type": "object",
                        "required": ["url"],
                        "properties": {
                          "url": { "type": "string", "format": "uri" },
                          "description": { "type": "string" },
                          "mimeType": { "type": "string" }
                        }
                      },
                      "accepted": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "scheme",
                          "network",
                          "asset",
                          "amount",
                          "payTo",
                          "maxTimeoutSeconds",
                          "extra"
                        ],
                        "properties": {
                          "scheme": { "type": "string", "enum": ["exact"] },
                          "network": {
                            "type": "string",
                            "enum": ["stellar:testnet", "stellar:pubnet"]
                          },
                          "asset": { "type": "string", "description": "SEP-41 token contract ID." },
                          "amount": {
                            "type": "string",
                            "pattern": "^[1-9]\\d*$",
                            "description": "Amount in the token's atomic units."
                          },
                          "payTo": { "type": "string" },
                          "maxTimeoutSeconds": { "type": "integer", "minimum": 1 },
                          "extra": { "type": "object", "additionalProperties": true }
                        }
                      },
                      "payload": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": ["transaction"],
                        "properties": {
                          "transaction": {
                            "type": "string",
                            "contentEncoding": "base64",
                            "description": "Wallet-signed Stellar transaction envelope XDR."
                          }
                        }
                      },
                      "extensions": { "type": "object", "additionalProperties": true }
                    }
                  },
                  "paymentRequirements": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": [
                      "scheme",
                      "network",
                      "asset",
                      "amount",
                      "payTo",
                      "maxTimeoutSeconds",
                      "extra"
                    ],
                    "properties": {
                      "scheme": { "type": "string", "enum": ["exact"] },
                      "network": {
                        "type": "string",
                        "enum": ["stellar:testnet", "stellar:pubnet"]
                      },
                      "asset": { "type": "string", "description": "SEP-41 token contract ID." },
                      "amount": {
                        "type": "string",
                        "pattern": "^[1-9]\\d*$",
                        "description": "Amount in the token's atomic units."
                      },
                      "payTo": { "type": "string" },
                      "maxTimeoutSeconds": { "type": "integer", "minimum": 1 },
                      "extra": { "type": "object", "additionalProperties": true }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["isValid"],
                  "properties": {
                    "isValid": { "type": "boolean" },
                    "invalidReason": { "type": "string" },
                    "invalidMessage": { "type": "string" },
                    "payer": { "type": "string" },
                    "extensions": { "type": "object", "additionalProperties": true },
                    "extensionResponses": { "type": "object", "additionalProperties": true },
                    "extra": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok", "error"],
                  "properties": {
                    "ok": { "type": "boolean", "enum": [false] },
                    "error": {
                      "type": "object",
                      "required": ["code", "message"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "details": { "type": "object", "additionalProperties": true }
                      }
                    },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/settle": {
      "post": {
        "operationId": "settlePayment",
        "summary": "Settle a previously verified exact Stellar x402 payment.",
        "tags": ["facilitator"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["x402Version", "paymentPayload", "paymentRequirements"],
                "properties": {
                  "x402Version": { "type": "integer", "enum": [2] },
                  "paymentPayload": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": ["x402Version", "accepted", "payload"],
                    "properties": {
                      "x402Version": { "type": "integer", "enum": [2] },
                      "resource": {
                        "type": "object",
                        "required": ["url"],
                        "properties": {
                          "url": { "type": "string", "format": "uri" },
                          "description": { "type": "string" },
                          "mimeType": { "type": "string" }
                        }
                      },
                      "accepted": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "scheme",
                          "network",
                          "asset",
                          "amount",
                          "payTo",
                          "maxTimeoutSeconds",
                          "extra"
                        ],
                        "properties": {
                          "scheme": { "type": "string", "enum": ["exact"] },
                          "network": {
                            "type": "string",
                            "enum": ["stellar:testnet", "stellar:pubnet"]
                          },
                          "asset": { "type": "string", "description": "SEP-41 token contract ID." },
                          "amount": {
                            "type": "string",
                            "pattern": "^[1-9]\\d*$",
                            "description": "Amount in the token's atomic units."
                          },
                          "payTo": { "type": "string" },
                          "maxTimeoutSeconds": { "type": "integer", "minimum": 1 },
                          "extra": { "type": "object", "additionalProperties": true }
                        }
                      },
                      "payload": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": ["transaction"],
                        "properties": {
                          "transaction": {
                            "type": "string",
                            "contentEncoding": "base64",
                            "description": "Wallet-signed Stellar transaction envelope XDR."
                          }
                        }
                      },
                      "extensions": { "type": "object", "additionalProperties": true }
                    }
                  },
                  "paymentRequirements": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": [
                      "scheme",
                      "network",
                      "asset",
                      "amount",
                      "payTo",
                      "maxTimeoutSeconds",
                      "extra"
                    ],
                    "properties": {
                      "scheme": { "type": "string", "enum": ["exact"] },
                      "network": {
                        "type": "string",
                        "enum": ["stellar:testnet", "stellar:pubnet"]
                      },
                      "asset": { "type": "string", "description": "SEP-41 token contract ID." },
                      "amount": {
                        "type": "string",
                        "pattern": "^[1-9]\\d*$",
                        "description": "Amount in the token's atomic units."
                      },
                      "payTo": { "type": "string" },
                      "maxTimeoutSeconds": { "type": "integer", "minimum": 1 },
                      "extra": { "type": "object", "additionalProperties": true }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "transaction", "network"],
                  "properties": {
                    "success": { "type": "boolean" },
                    "errorReason": { "type": "string" },
                    "errorMessage": { "type": "string" },
                    "payer": { "type": "string" },
                    "transaction": { "type": "string" },
                    "network": { "type": "string", "enum": ["stellar:testnet", "stellar:pubnet"] },
                    "amount": { "type": "string", "pattern": "^[1-9]\\d*$" },
                    "extensions": { "type": "object", "additionalProperties": true },
                    "extensionResponses": { "type": "object", "additionalProperties": true },
                    "extra": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok", "error"],
                  "properties": {
                    "ok": { "type": "boolean", "enum": [false] },
                    "error": {
                      "type": "object",
                      "required": ["code", "message"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "details": { "type": "object", "additionalProperties": true }
                      }
                    },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/receipts/{receiptId}": {
      "get": {
        "operationId": "getReceipt",
        "summary": "Fetch a settlement receipt.",
        "tags": ["facilitator"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "type": "object", "additionalProperties": true } }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok", "error"],
                  "properties": {
                    "ok": { "type": "boolean", "enum": [false] },
                    "error": {
                      "type": "object",
                      "required": ["code", "message"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "details": { "type": "object", "additionalProperties": true }
                      }
                    },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-sessions": {
      "post": {
        "operationId": "createPaymentSession",
        "summary": "Create a capped Stellar upto payment session.",
        "tags": ["facilitator"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "scheme",
                  "network",
                  "buyer",
                  "asset",
                  "capAmount",
                  "payTo",
                  "expiresAtLedger"
                ],
                "properties": {
                  "scheme": { "type": "string", "enum": ["upto"] },
                  "network": { "type": "string", "enum": ["stellar:testnet", "stellar:pubnet"] },
                  "buyer": { "type": "string" },
                  "asset": {
                    "type": "object",
                    "required": ["code", "issuer"],
                    "properties": { "code": { "type": "string" }, "issuer": { "type": "string" } }
                  },
                  "capAmount": { "type": "string" },
                  "payTo": { "type": "string" },
                  "resourceId": { "type": "string" },
                  "sellerId": { "type": "string" },
                  "expiresAtLedger": { "type": "integer" },
                  "currentLedger": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "network",
                    "buyer",
                    "payTo",
                    "assetCode",
                    "assetIssuer",
                    "assetContractId",
                    "capAmount",
                    "spentAmount",
                    "remainingAmount",
                    "contractId",
                    "contractSessionId",
                    "resourceHash",
                    "expiresAtLedger",
                    "status"
                  ],
                  "properties": {
                    "id": { "type": "string" },
                    "resourceId": { "type": ["string", "null"] },
                    "sellerId": { "type": ["string", "null"] },
                    "network": { "type": "string", "enum": ["stellar:testnet", "stellar:pubnet"] },
                    "buyer": { "type": "string" },
                    "payTo": { "type": "string" },
                    "assetCode": { "type": "string" },
                    "assetIssuer": { "type": "string" },
                    "assetContractId": { "type": "string" },
                    "capAmount": { "type": "string" },
                    "spentAmount": { "type": "string" },
                    "remainingAmount": { "type": "string" },
                    "contractId": { "type": "string" },
                    "contractSessionId": { "type": "string" },
                    "resourceHash": { "type": "string" },
                    "expiresAtLedger": { "type": "integer" },
                    "status": { "type": "string", "enum": ["open", "settled"] },
                    "transactionHash": { "type": ["string", "null"] },
                    "ledger": { "type": ["integer", "null"] },
                    "usageHash": { "type": ["string", "null"] },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "updatedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok", "error"],
                  "properties": {
                    "ok": { "type": "boolean", "enum": [false] },
                    "error": {
                      "type": "object",
                      "required": ["code", "message"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "details": { "type": "object", "additionalProperties": true }
                      }
                    },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}